Fix execute_ceph_cmd function race condition
The function execute_ceph_cmd in ceph_common.sh uses a temporary file
to get the command output. It uses the same filename for all the
calls. This is not thread-safe and since many processes are calling
this function, a race condition can happen.

The filename now has the PID concatenated at the end. This way the
filename will never be the same for processes calling this function at
the same time.

The filename was '.ceph_cmd_out' and has changed to
'._ceph_cmd_out_$$', where $$ will be replaced by the PID

Test-Plan:
    PASS: Call the script and check if the file was created using the
          new filename template and it is removed when the function
          returns

Closes-bug: https://bugs.launchpad.net/starlingx/+bug/2085648

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>